requirements.txt
qiskit-aqua>=0.4.2
numpy>=1.13
revkit>=3.1
